requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-rest-authtoken==2.1.4
djangorestframework==3.15.1
et-xmlfile==1.1.0
Markdown==3.6
openpyxl==3.1.2
sqlparse==0.5.0
tzdata==2024.1
